[ALPS09897705] mcupm: mt6993: coverity fix
[Description]
coverity fix

[Test]
Boot Test, mcupm_alive

MTK-Commit-Id: 39e9dbfa1a5f24da1049e4d4f27835e35b257819

CR-Id: ALPS09897705
Feature: mcupm_driver

+++ drivers/misc/mediatek/mcupm/v3/common/mcupm_timesync.c
@@ -83,7 +83,7 @@ static void mcupm_ts_update(int suspended, u64 tick, u64 ts)
 	mcupm_base_ver = (mcupm_base_ver + 1)%(timesync_max_ver+1);
 
 	/* make header: freeze and version */
-	header = suspended ? TIMESYNC_HEADER_FREEZE : 0;
+	header = suspended ? (u32)TIMESYNC_HEADER_FREEZE : 0;
 
 	header |= ((mcupm_base_ver << timesync_header_ver_ofs) &
 		timesync_header_ver_mask);
